requirements.txt
--pre
torch
tensordict-nightly
torchrl-nightly
einops
